Refactor HomeScreen and related components: integrate HomeProvider, remove useHome hook, and implement HomeView for better state management
