+++ util/currency.ts
@@ -1,4 +1,6 @@
 import {
+  ICurrency,
+  IDestinationCurrencyType,
   TradingPair,
   TradingPairsByQuoteCurrency,
 } from "@/types/currencies.type";
@@ -14,4 +16,26 @@ export const groupByQuoteCurrency = (
     acc[quoteCurrency].push(pair);
     return acc;
   }, {} as TradingPairsByQuoteCurrency);
+};
+
+export const filterCurrenciesByTradingPairs = (
+  currencies: ICurrency[],
+  tradingPairs: { base_currency: string; id: string }[]
+) => {
+  if (!currencies || !tradingPairs) return [];
+
+  // Create a map of currency IDs to their trading pair IDs
+  const currencyToPairMap = new Map<string, string>();
+
+  tradingPairs.forEach((pair) => {
+    currencyToPairMap.set(pair.base_currency, pair.id);
+  });
+
+  // Filter currencies and add trading pair ID as market
+  return currencies
+    .filter((currency) => currencyToPairMap.has(currency.id))
+    .map((currency) => ({
+      ...currency,
+      market: currencyToPairMap.get(currency.id) || "",
+    })) as IDestinationCurrencyType[];
 };
